+++ framework/components/blockchain/blockchain.go
@@ -5,6 +5,7 @@ import (
 	"github.com/smartcontractkit/chainlink-testing-framework/framework"
 )
 
+// Input is a blockchain network configuration params
 type Input struct {
 	Type                     string   `toml:"type" validate:"required,oneof=anvil geth"`
 	Image                    string   `toml:"image" validate:"required"`
@@ -15,11 +16,13 @@ type Input struct {
 	Out                      *Output  `toml:"out"`
 }
 
+// Output is a blockchain network output, ChainID and one or more nodes that forms the network
 type Output struct {
 	ChainID string  `toml:"chain_id" validate:"required"`
 	Nodes   []*Node `toml:"nodes" validate:"required"`
 }
 
+// Node represents blockchain node output, URLs required for connection locally and inside docker network
 type Node struct {
 	WSUrl                 string `toml:"ws_url"`
 	HTTPUrl               string `toml:"http_url"`
@@ -27,6 +30,9 @@ type Node struct {
 	DockerInternalHTTPUrl string `toml:"docker_internal_http_url" validate:"required"`
 }
 
+// NewBlockchainNetwork this is an abstraction that can spin up various blockchain network simulators
+// - Anvil
+// - Geth
 func NewBlockchainNetwork(input *Input) (*Output, error) {
 	if input.Out != nil && framework.UseCache() {
 		return input.Out, nil
